+++ pinot-tools/src/main/java/org/apache/pinot/tools/admin/PinotAdministrator.java
@@ -29,11 +29,10 @@ import org.apache.pinot.tools.admin.command.AvroSchemaToPinotSchema;
 import org.apache.pinot.tools.admin.command.BackfillDateTimeColumnCommand;
 import org.apache.pinot.tools.admin.command.ChangeNumReplicasCommand;
 import org.apache.pinot.tools.admin.command.ChangeTableState;
+import org.apache.pinot.tools.admin.command.OperateClusterConfigCommand;
 import org.apache.pinot.tools.admin.command.CreateSegmentCommand;
 import org.apache.pinot.tools.admin.command.DeleteClusterCommand;
-import org.apache.pinot.tools.admin.command.DeleteClusterConfigCommand;
 import org.apache.pinot.tools.admin.command.GenerateDataCommand;
-import org.apache.pinot.tools.admin.command.GetClusterConfigsCommand;
 import org.apache.pinot.tools.admin.command.LaunchDataIngestionJobCommand;
 import org.apache.pinot.tools.admin.command.MoveReplicaGroup;
 import org.apache.pinot.tools.admin.command.OfflineSegmentIntervalCheckerCommand;
@@ -49,7 +48,6 @@ import org.apache.pinot.tools.admin.command.StartServerCommand;
 import org.apache.pinot.tools.admin.command.StartZookeeperCommand;
 import org.apache.pinot.tools.admin.command.StopProcessCommand;
 import org.apache.pinot.tools.admin.command.StreamAvroIntoKafkaCommand;
-import org.apache.pinot.tools.admin.command.UpdateClusterConfigCommand;
 import org.apache.pinot.tools.admin.command.UploadSegmentCommand;
 import org.apache.pinot.tools.admin.command.ValidateConfigCommand;
 import org.apache.pinot.tools.admin.command.VerifyClusterStateCommand;
@@ -92,9 +90,7 @@ public class PinotAdministrator {
   @Argument(handler = SubCommandHandler.class, metaVar = "<subCommand>")
   @SubCommands({
       @SubCommand(name = "QuickStart", impl = QuickStartCommand.class),
-      @SubCommand(name = "GetClusterConfigs", impl = GetClusterConfigsCommand.class),
-      @SubCommand(name = "UpdateClusterConfig", impl = UpdateClusterConfigCommand.class),
-      @SubCommand(name = "DeleteClusterConfig", impl = DeleteClusterConfigCommand.class),
+      @SubCommand(name = "OperateClusterConfig", impl = OperateClusterConfigCommand.class),
       @SubCommand(name = "GenerateData", impl = GenerateDataCommand.class),
       @SubCommand(name = "LaunchDataIngestionJob", impl = LaunchDataIngestionJobCommand.class),
       @SubCommand(name = "CreateSegment", impl = CreateSegmentCommand.class),
